Type GameUI_update scroll-dispatch section index as u8
Declare r29v as u8 so the lbl_803DD8B7 store emits a raw stb (no clrlwi
mask), matching retail. GameUI_update 99.49 -> 99.64.

+++ src/main/dll/dll_0000_gameui.c
@@ -4098,7 +4098,7 @@ void GameUI_update(void)
     s16 cx;
     u8 f25 = 1;
     u8 f26 = 0;
-    int r29v;
+    u8 r29v;
     int flags;
 
     gCMenuButtons = getButtonsJustPressed(0);
